static/requirements.txt
blinker==1.9.0
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
Flask==3.0.3
Flask-Mail==0.10.0
Flask-CORS==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1

pdfminer.six==20231228
pdfplumber==0.11.4
pillow==12.2.0
pycparser==3.0
pypdfium2==5.6.0
python-docx==1.1.2
typing_extensions==4.15.0
Werkzeug==3.0.3
gunicorn==22.0.0
bcrypt==4.1.2
